fix: rename dsmComponents to dsm
