+++ src/components/dsm/Announcements.js
