+++ src/components/routes/Home.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Box } from '@mui/material';
 
-import DSMHeader from '../dsmComponents/DSMHeader';
-import DSMBody from '../dsmComponents/DSMBody';
+import DSMHeader from '../dsm/DSMHeader';
+import DSMBody from '../dsm/DSMBody';
 
 export default function HomeContainer() {
   return (
